+++ magiccube/__init__.py
@@ -1,3 +1,11 @@
-__all__ = ["cube"]
+"""MagicCube: A fast implementation of the Rubik Cube based in Python 3.x."""
+
+__all__ = ["cube", "cube_base", "cube_piece",
+           "cube_move", "solver"]
 
 from .cube import Cube
+from .cube_base import Color, CubeException, Face, PieceType
+from .cube_piece import CubePiece
+from .cube_move import CubeMove, CubeMoveType
+from .cube_print import CubePrintStr, Terminal
+from .solver.basic.basic_solver import BasicSolver